What I did today: Raghad
Converted the for loop to a filter in the __init__ function.

Created a function called is_one_letter_diff to be used in the valid_transformations function.

Created a new function that is a combination of the Fisher-Yates shuffle and the card list of 28 letters.

Fixed the star_card function and added a new else statement: if the player doesn’t get the star card, they will receive a useful letter instead.

Created a function to generate a stack of 40 cards.

Used a different quicksort algorithm from GitHub.

Also fixed the word generator and the coin function.